Move wallet stuff into wallet/
