+++ frontend/src/efi-ui/wallet/ConnectWalletDialog/ConnectWalletDialog.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from "react";
 import SimpleDialog from "src/efi-ui/base/Dialog/Dialog";
-import { ConnectWalletButtons } from "src/efi-ui/ConnectWalletButtons/ConnectWalletButtons";
+import { ConnectWalletButtons } from "src/efi-ui/wallet/ConnectWalletButtons/ConnectWalletButtons";
 
 interface ConnectWalletDialogProps {
   isOpen: boolean;
