+++ frontend/src/efi-ui/wallet/WalletJazzicon/WalletJazzicon.tsx
